[CLEANUP] Refatoração e melhorias na classe Statup
- implementação para que o IConfiguration utilize as configurações de acordo com o ambiente
- separando as configurações do Identity, Swagger e Api em classes separadas
