+++ src/services/Estudos.NSE.Identidade.API/Configuration/ApiConfig.cs
@@ -1,0 +1,36 @@
+﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
+namespace Estudos.NSE.Identidade.API.Configuration
+{
+    public static class ApiConfig
+    {
+        public static void AddApiConfiguration(this IServiceCollection services)
+        {
+            services.AddControllers();
+
+        }
+
+        public static void UseApiConfiguration(this IApplicationBuilder app, IWebHostEnvironment env)
+        {
+            if (env.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+            }
+
+            app.UseHttpsRedirection();
+
+            app.UseRouting();
+
+            app.UseIdentityConfiguration();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
+
+        }
+    }
+}
