+++ src/services/Estudos.NSE.Identidade.API/Configuration/IdentityConfig.cs
@@ -1,0 +1,87 @@
+﻿using System.Text;
+using Estudos.NSE.Identidade.API.Data;
+using Estudos.NSE.Identidade.API.Extensions;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.IdentityModel.Tokens;
+
+namespace Estudos.NSE.Identidade.API.Configuration
+{
+    public static class IdentityConfig
+    {
+        public static void AddIdentityConfiguration(this IServiceCollection services, IConfiguration configuration)
+        {
+            services.AddDbContext<ApplicationDbContext>(opt =>
+            {
+                opt.UseSqlServer(configuration.GetConnectionString("DefaultConnection"));
+            });
+
+            services.AddDefaultIdentity<IdentityUser>(options =>
+                {
+                    //senha
+                    options.Password.RequiredLength = 8;
+                    options.Password.RequireLowercase = false;
+                    options.Password.RequireUppercase = false;
+                    options.Password.RequireNonAlphanumeric = false;
+                    options.Password.RequireDigit = false;
+
+                    //confirmação de informãções
+                    options.SignIn.RequireConfirmedEmail = false;
+                    options.SignIn.RequireConfirmedPhoneNumber = false;
+                })
+                .AddRoles<IdentityRole>()
+                .AddErrorDescriber<IdentityMensagensPortugues>()
+                .AddEntityFrameworkStores<ApplicationDbContext>()
+                .AddDefaultTokenProviders();
+
+
+            #region JWT
+            var appSettingsSection = configuration.GetSection("AppSettings");
+            services.Configure<AppSettings>(appSettingsSection);
+
+            var appSettings = appSettingsSection.Get<AppSettings>();
+            var key = Encoding.ASCII.GetBytes(appSettings.Secret);
+
+            services.AddAuthentication(opt =>
+            {
+                opt.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+                opt.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+            }).AddJwtBearer(bearerOptions =>
+            {
+                bearerOptions.RequireHttpsMetadata = true;
+                bearerOptions.SaveToken = true;
+
+                //parametros de validação do token
+                bearerOptions.TokenValidationParameters = new TokenValidationParameters
+                {
+                    //validar a assinatura do token?
+                    ValidateIssuerSigningKey = true,
+                    //chave de criptografia
+                    IssuerSigningKey = new SymmetricSecurityKey(key),
+                    //validar o emisor?
+                    ValidateIssuer = true,
+                    //validar os dominios onde o token é válido?
+                    ValidateAudience = true,
+                    //seta o doimio onde o token é valido
+                    ValidAudience = appSettings.ValidoEm,
+                    //seta o emissor
+                    ValidIssuer = appSettings.Emissor
+                };
+            });
+            #endregion
+
+
+        }
+
+        public static void UseIdentityConfiguration(this IApplicationBuilder app)
+        {
+            app.UseAuthentication();
+            app.UseAuthorization();
+        }
+
+    }
+}
